backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-google-genai
langchain-tavily
pydantic
pydantic-settings
tavily-python
axios
cors
python-multipart
langchain-community
langchain-core
groq
langchain_groq
sqlalchemy

